Remove cric_home app and migrate its functionality to cric_users
