--- cric_home/apps.py
@@ -1,6 +1,0 @@
-from django.apps import AppConfig
-
-
-class CricHomeConfig(AppConfig):
-    default_auto_field = "django.db.models.BigAutoField"
-    name = "cric_home"
