+++ cric_users/urls.py
@@ -2,7 +2,9 @@ from django.urls import path
 from . import views
 
 urlpatterns = [
-    path('', views.profile_view, name="profile"),
+    path('', views.home_view, name="home"),
+    path('match/<int:match_id>/', views.match_detail_view, name='match_detail'),
+    path('profile/', views.profile_view, name="profile"),
     path('edit/', views.profile_edit_view, name="profile-edit"),
     path('onboarding/', views.profile_edit_view, name="profile-onboarding"),
     path('settings/', views.profile_settings_view, name="profile-settings"),
